+++ g2p_registrant/wizard/__init__.py
@@ -5,7 +5,4 @@
 # All Rights Reserved.
 #################################################################################
 
-from . import models
-from . import wizard
-#from . import controllers
-
+from . import disable_registrant